+++ TonoJit/JitStage.cs
@@ -12,20 +12,8 @@ namespace Tono.Jit
     /// 工程やワーク全体を動かす 根幹のオブジェクト
     /// </summary>
     [JacTarget(Name = "Stage")]
-    public partial class JitStage : JitVariable, IJitObjectID
+    public partial class JitStage : JitStageSubset
     {
-        public string ID { get; set; } = JacInterpreter.MakeID("Stage");
-
-        /// <summary>
-        /// Stage runtime data
-        /// </summary>
-        public IJitStageEngine Engine { get; set; }
-
-        /// <summary>
-        /// Jit Sub Model
-        /// </summary>
-        public IJitStageModel Model { get; set; }
-
         /// <summary>
         /// the constructor of this class
         /// </summary>
@@ -34,43 +22,6 @@ namespace Tono.Jit
             Classes.Add(":Stage");
             Engine = new JitStageEngine();
             Model = new JitStageModel();
-        }
-
-        public override int GetHashCode()
-        {
-            return ID.GetHashCode();
-        }
-        public override bool Equals(object obj)
-        {
-            if (obj is JitStage st)
-            {
-                return st.ID == ID;
-            }
-            else
-            {
-                return false;
-            }
-        }
-        public override string ToString()
-        {
-            return $"{GetType().Name} ID={ID}";
-        }
-
-        [JacListAdd(PropertyName = "Procs")]
-        public void ProcsAdd(object obj)
-        {
-            Model.ProcsAdd(obj);
-        }
-
-        [JacListRemove(PropertyName = "Procs")]
-        public void ProcsRemove(object obj)
-        {
-            Model.ProcsRemove(obj);
-        }
-        [JacListAdd(PropertyName = "ProcLinks")]
-        public void AddProcLinks(object description)
-        {
-            Model.AddProcLinks(description);
         }
     }
 }
