+++ TonoJit/JitStageModel.cs
@@ -30,30 +30,6 @@ namespace Tono.Jit
             Procs = new ProcessSet();
         }
 
-        public void ProcsAdd(object obj)
-        {
-            if (obj is JitProcess proc)
-            {
-                Procs.Add(proc);
-            }
-            else
-            {
-                throw new JacException(JacException.Codes.TypeMismatch, $"Procs.Add type mismatch arg type={(obj?.GetType().Name ?? "null")}");
-            }
-        }
-
-        public void ProcsRemove(object obj)
-        {
-            if (obj is JitProcess proc)
-            {
-                Procs.Remove(proc);
-            }
-            else
-            {
-                throw new JacException(JacException.Codes.TypeMismatch, $"Procs.Add type mismatch arg type={(obj?.GetType().Name ?? "null")}");
-            }
-        }
-
         public JitProcess FindProcess(string processKey, bool isReturnNull = false)
         {
             if (string.IsNullOrEmpty(processKey))
@@ -73,24 +49,6 @@ namespace Tono.Jit
                 throw new JitException(JitException.FormatNoProcKey, processKey);
             }
             return ret;
-        }
-
-        public void AddProcLinks(object description)
-        {
-            if (description == null)
-            {
-                throw new JitException(JitException.NullValue, "ProcLinks");
-            }
-            if (description is JacPushLinkDescription push)
-            {
-                var key1 = push.From is JitProcess p1 ? p1.ID : push.From?.ToString();
-                var key2 = push.To is JitProcess p2 ? p2.ID : push.To?.ToString();
-                AddProcessLink(key1, key2);
-            }
-            else
-            {
-                throw new JitException(JitException.TypeMissmatch, $"Expecting type JacPushLinkDescription but {description.GetType().Name}");
-            }
         }
 
         /// <summary>
